requirements.txt
langchain_groq
crewai
crewai_tools
langchain_community
duckduckgo-search
pandas
yfinance
streamlit
plotly